# Store Chow-Liu structure in lists while it is being built